fix(clipboard): route every renderer clipboard write through the main process
Electron 44 removed the clipboard module from the renderer entirely, so the
three remaining writeText call sites threw at runtime while still typechecking
clean. Move both text and image writes behind IPC helpers.
