+++ views/components/tab-area/plugin-wrapper.tsx
@@ -3,9 +3,9 @@ import type { Plugin } from 'views/services/plugin-manager'
 
 import { Card, TextArea, Button, Intent } from '@blueprintjs/core'
 import * as Sentry from '@sentry/electron'
-import { clipboard } from 'electron'
 import React, { Component } from 'react'
 import { useTranslation } from 'react-i18next'
+import { writeClipboardText } from 'views/services/clipboard'
 
 interface Props {
   plugin: Plugin
@@ -49,7 +49,7 @@ class ErrorBoundary extends Component<BoundaryProps, State> {
   handleCopy = (): void => {
     const { error, info } = this.state
     if (!error || !info) return
-    clipboard.writeText([error.stack, info.componentStack].join('\n'))
+    void writeClipboardText([error.stack, info.componentStack].join('\n'))
   }
 
   componentDidMount(): void {
