+++ views/services.ts
@@ -9,6 +9,7 @@ import { dbg } from 'views/env-parts/dbg'
 import i18next from 'views/env-parts/i18next'
 import { toggleModal } from 'views/env-parts/modal'
 import { log, error } from 'views/services/alert'
+import { writeClipboardText } from 'views/services/clipboard'
 import { isInGame } from 'views/utils/game-utils'
 
 const gameAPIBroadcaster: GameAPIBroadcaster = remote.require('./lib/game-api-broadcaster')
@@ -127,7 +128,7 @@ class GameResponse {
     })
     Object.defineProperty(this, 'ClickToCopy -->', {
       get: () => {
-        require('electron').clipboard.writeText(JSON.stringify({ path, body, postBody }))
+        void writeClipboardText(JSON.stringify({ path, body, postBody }))
         return `Copied: ${this.path}`
       },
     })
